==> dredFISH/Design/Misc/scPNMF/00.prep_downloaded_data_smrt.ipynb <==
import time
import numpy as np
from scipy import sparse
import pandas as pd
import h5py
import anndata

from dredFISH.Analysis import basicu

res_dir = '/bigstore/GeneralStorage/fangming/projects/dredfish/data/rna/'

# ddir = '/bigstore/binfo/mouse/Brain/Sequencing/Allen_10X_CtxHippo/source/'
ddir = '/bigstore/binfo/mouse/Brain/Sequencing/Allen_SmartSeq_CtxHippo/source/'

# load metadata

f = ddir + 'metadata.csv'
df1 = pd.read_csv(f, low_memory=False).set_index('sample_name')
f = ddir + 'tsne.csv'
df2 = pd.read_csv(f, low_memory=False).set_index('sample_name')
df = df1.join(df2) #.astype({'cluster_order': str})

# load data matrix 
Unpack the `hdf5` file:
- `data` 
    - `exon` (csc_matrix)
    - `t_exon` (csc_matrix)
    - `total_exon_counts` (1d array)
- `gene_names`  (1d array)
- `sample_names`  (1d array)


csc_matrix: dims, i, p, x
scipy
`csc_matrix((data, indices, indptr), [shape=(M, N)])`

ti = time.time()
f = ddir + 'expression_matrix.hdf5'
with h5py.File(f, 'r') as fh:
    print(fh.keys())
    print(fh['data'].keys())
    print(fh['data']['exon'].keys())
    
    dat = fh['data']['exon']
    dims = dat['dims'][()]
    print(dims)

    i = dat['i'][()]
    p = dat['p'][()]
    x = dat['x'][()]
    mat = sparse.csc_matrix((x,i,p), shape=dims)
    print(mat.shape)
    
    genes = fh['gene_names'][()]
    genes = np.array([gene.decode() for gene in genes])
    
    cells = fh['sample_names'][()] 
    cells = np.array([cell.decode() for cell in cells])

cells.shape, df.shape, np.intersect1d(cells, df.index.values).shape

# for each cell in meta, where it is in the matrix
cell_idx = basicu.get_index_from_array(cells, df.index.values)
(cell_idx == -1).sum()

adata = anndata.AnnData(mat.tocsr()[cell_idx,:],  # mat -- csr: easy for row(cell) selection
                   var=pd.DataFrame(index=genes), 
                   obs=df)
adata

dataid = 'scrna_ss_ctxhippo_a'
output = f'{res_dir}{dataid}_exon_count_matrix_v5.h5ad'
print(output)
adata.write(output)
!chmod 444 $output

# Done and check

# file = f'{res_dir}{dataid}_exon_count_matrix_v5.h5ad'
file = output
adata2 = anndata.read(file, backed='r')
adata2
