==> dredFISH/Design/Misc/scPNMF/01.prep_data_forR.ipynb <==
import time
import os 
import numpy as np
from scipy import sparse
import pandas as pd
import h5py
import anndata


from dredFISH.Analysis import basicu

ddir = '/bigstore/GeneralStorage/fangming/projects/dredfish/data/rna/'
res_dir = '/bigstore/GeneralStorage/fangming/projects/dredfish/res_dpnmf/'

%%time
# read in data
f = os.path.join(ddir, 'scrna_ss_ctxhippo_a_exon_count_matrix_v5.h5ad')
adata = anndata.read(f,
                     # backed='r',
                    )
adata

# get the NN genes
f = '/bigstore/binfo/mouse/Brain/DRedFISH/Allen_V3_Reference/10X/genes.npy' 
tenx_genes = np.load(f)
f = '/bigstore/binfo/mouse/Brain/DRedFISH/Allen_V3_Reference/SmartSeq/genes.npy' 
smrt_genes = np.load(f)
assert np.all(tenx_genes == smrt_genes)
nn_genes = tenx_genes
nn_genes.shape

# prep
- select relevant genes only
- subsample cells
- dense matrix
- numpy format


nn_genes_idx = basicu.get_index_from_array(adata.var.index.values, nn_genes)
(nn_genes_idx == -1).sum()

%%time
adata_nng = adata[:,adata.var.iloc[nn_genes_idx].index].copy()
adata_nng

# all cells
# randomly select cells L5 (allow oversample)

# randomly select cells L3 (does not allow oversamp)
clstcol = 'subclass_label'
n = 100
dfsub, idxsub = basicu.stratified_sample(adata_nng.obs, clstcol, n, 
                                         return_idx=True, group_keys=False, sort=False, 
                                         random_state=0, 
                                        )

adatasub = adata_nng[adata_nng.obs.iloc[idxsub].index].copy()
adatasub

adatasub.obs.groupby(clstcol).size()

# convert to dense and format it; need the matrix and the labels
X = adatasub.X.toarray().T # gene by cell
y = adatasub.obs[clstcol].values
X.shape, y.shape
y

out_X = os.path.join(res_dir, 'test_X.npy') 
out_y = os.path.join(res_dir, 'test_y.npy') 
np.save(out_X, X)
np.save(out_y, y)

print(out_X)
